% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-guides.R
\name{one_line_style}
\alias{one_line_style}
\title{Put all code on one line}
\usage{
one_line_style()
}
\description{
This is for real programmers. All code on one line. No comments. I am
serious. Use this style guide in conjunction with the package styler to
format your code like nerds do.
}
\examples{
style_text(
  "sum <- function(x, y) {
   x + # hi
   y
  }
  1 + 1",
  style = one_line_style
)
}
