% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneliner.R
\docType{package}
\name{oneliner-package}
\alias{oneliner}
\alias{oneliner-package}
\title{We all love one-liners}
\description{
This package uses the styler package and writes a style guide
that puts all your code on one line. I claim this to be the
most useful programming tool invented.
}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

}
