% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-is.R
\name{pd_is}
\alias{pd_is}
\alias{is_brace_expr}
\alias{is_curly_brace_expr}
\alias{is_round_brace_expr}
\title{Check whether a parse table corresponds to a certain expression}
\usage{
is_brace_expr(pd, type = c("'('", "'{'", "'['"))

is_curly_brace_expr(...)

is_round_brace_expr(...)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Check whether a parse table corresponds to a certain expression
}
\section{Functions}{
\itemize{
\item \code{is_brace_expr}: Checks whether `pd` contains an expression wrapped in
braces.

\item \code{is_curly_brace_expr}: Checks whether `pd` contains an expression wrapped in
curly brackets.

\item \code{is_round_brace_expr}: Checks whether `pd` contains an expression wrapped in
round brackets.
}}

\keyword{internal}
