% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler-io.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{style_pkg}
\alias{reexports}
\alias{style_file}
\alias{reexports}
\alias{style_text}
\alias{reexports}
\alias{style_dir}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{styler}{\code{\link[styler]{style_pkg}}, \code{\link[styler]{style_file}}, \code{\link[styler]{style_text}}, \code{\link[styler]{style_dir}}}
}}

